Add files via upload
This is a Coursera Dataset taken from Applied Machine Learning in Python It is training and testing datasets of educational video features from the VLE Dataset put together by researcher Sahan Bulathwela at University College London.   It is provided with two data files for use in training and validating your models: train.csv and test.csv. Each row in these two files corresponds to a single educational video and includes information about the diverse properties of the video content. According to the dataset, The target variable is `engagement` which was defined as True if the median percentage of the video watched across all viewers was at least 30%, and False otherwise.